+++ app/models/barbeque/docker_container.rb
@@ -1,0 +1,2 @@
+class Barbeque::DockerContainer < Barbeque::ApplicationRecord
+end