+++ app/models/barbeque/ecs_hako_task.rb
@@ -1,0 +1,2 @@
+class Barbeque::EcsHakoTask < Barbeque::ApplicationRecord
+end