+++ lib/barbeque/execution_poller.rb
@@ -1,0 +1,34 @@
+module Barbeque
+  class ExecutionPoller
+    def initialize
+      @stop_requested = false
+    end
+
+    def run
+      Barbeque::JobExecution.running.find_in_batches do |job_executions|
+        job_executions.shuffle.each do |job_execution|
+          if @stop_requested
+            return
+          end
+          job_execution.with_lock do
+            if job_execution.running?
+              poll(job_execution)
+            end
+          end
+        end
+      end
+      sleep 1
+    end
+
+    def stop
+      @stop_requested = true
+    end
+
+    private
+
+    def poll(job_execution)
+      executor = Executor.create
+      executor.poll_execution(job_execution)
+    end
+  end
+end