+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 ## v1.0.0 (xxxx-xx-xx)
+- Introduce Executor as a replacement of Runner
+  - `runner` and `runner_options` is renamed to `executor` and `executor_options` respectively
+  - Now `rake barbeque:worker` launches three types of process
+    - Runner: receives message from SQS queue, starts job execution and stores its identifier to the database
+      - In Executor::Docker, the identifier is container id
+      - In Executor::Hako, the identifier is ECS cluster and task ARN
+    - ExecutionPoller: polls execution status and reflect it to the database
+      - In Executor::Docker, uses `docker inspect` command
+      - In Executor::Docker, uses S3 task notification JSON
+    - RetryPoller: polls retry status and reflect it to the database
+      - Same with ExecutionPoller
+  - Add `maximum_concurrent_executions` configuration to config/barbeque.yml
+    - It controls the number of concurrent job executions
+    - The limit is disabled by default
 - Add `sqs_receive_message_wait_time` configuration to config/barbeque.yml
   - This option controls ReceiveMessageWaitTimeSeconds attribute of SQS queue
   - The default value is changed from 20s to 10s